pyproject.toml
# SPDX-FileCopyrightText: 2019-2020 Magenta ApS
# SPDX-License-Identifier: MPL-2.0
[tool.poetry]
name = "adguidsync"
version = "0.1.0"
description = "OS2mo AMQP Trigger receiver that writes ADGUID to an IT system"
authors = ["Magenta ApS <info@magenta.dk>"]
license = "MPL-2.0"
readme = "README.md"
homepage = "https://magenta.dk/"
repository = "https://git.magenta.dk/rammearkitektur/os2mo-triggers/os2mo-amqp-trigger-adguid-sync"
keywords = ["os2mo", "amqp"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
fastramqpi = "^0.0.0"
fastapi = "^0.78.0"
gql = "^3.3.0"
more-itertools = "^8.13.0"
strawberry-graphql = "^0.116.4"
raclients = "^1.2.5"
rautils = "^1.0.10"
ldap3 = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"
reorder-python-imports = "^3.1.0"
mypy = "^0.950"
pre-commit = "^2.19.0"
pylint-pydantic = "^0.1.4"
requests = "^2.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode="auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = "True"
warn_unused_configs = "True"
check_untyped_defs = "True"
disallow_untyped_defs = "True"
plugins = "pydantic.mypy"

[tool.flake8]
max-line-length = 88
exclude = [
".git", "__pycache__", ".mypy_cache", ".pytest_cache", ".venv", ".gitlab"
]
